# Improve `cupy.argwhere` and `cupy.nonzero`