--- a/cupy/core/_routines_math.pxd
+++ b/cupy/core/_routines_math.pxd
@@ -19,7 +19,8 @@ cpdef enum scan_op:
     SCAN_SUM = 0
     SCAN_PROD = 1
 
-cdef ndarray scan(ndarray a, op, dtype=*, ndarray out=*)
+cdef ndarray scan(ndarray a, op, dtype=*, ndarray out=*, incomplete=*,
+                  block_size=*)
 cdef object _sum_auto_dtype
 cdef object _add
 cdef object _conj