This PR improves `_ndarray_argwhere` in cupy/core/_routines_indexing.pyx, which is the core routine for `cupy.argwhere` and `cupy.nonzero`.

This ~depends on https://github.com/cupy/cupy/pull/4366 and~ is related to https://github.com/cupy/cupy/pull/4305.

Below is a performance comparison of `argwhere`. (updated results as there was a bug in the time measurement script)

shape | density | NumPy | CuPy (master) | CuPy (#4305) | CuPy (this PR, https://github.com/cupy/cupy/pull/4367/commits/34f942248a1ac1aa72d3b3b9e6d2b5478162a547)
------|---------|-------|-----|-----|-----
(1000000,) | 0.1 | 1.595 ms | 0.116 ms | 0.127 ms | 0.131 ms
(1000000,) | 0.5 | 4.062 ms | 0.117 ms | 0.129 ms | 0.128 ms
(1000, 1000) | 0.1 | 3.911 ms | 0.120 ms | 0.133 ms | 0.136 ms
(1000, 1000) | 0.5 | 9.024 ms | 0.120 ms | 0.144 ms | 0.140 ms
(10000000,) | 0.1 | 16.452 ms | 0.584 ms | 0.470 ms | 0.270 ms
(10000000,) | 0.5 | 46.937 ms | 0.599 ms | 0.494 ms | 0.281 ms
(3162, 3163) | 0.1 | 37.605 ms | 0.589 ms | 0.531 ms | 0.309 ms
(3162, 3163) | 0.5 | 107.924 ms | 0.682 ms | 0.599 ms | 0.365 ms
(100000000,) | 0.1 | 187.557 ms | 5.350 ms | 4.128 ms | 1.998 ms
(100000000,) | 0.5 | 492.996 ms | 5.483 ms | 4.446 ms | 2.122 ms
(10000, 10000) | 0.1 | 404.387 ms | 5.365 ms | 4.691 ms | 2.383 ms
(10000, 10000) | 0.5 | 1068.958 ms | 6.229 ms | 5.418 ms | 2.916 ms

<details>
<summary> Click to see the script </summary>

```python

import math
import numpy
import cupy
from cupy import testing
from cupyx.time import repeat

routine_list = ('argwhere', )
# routine_list = ('argwhere', 'nonzero')

size_list = (1000000, 10000000, 100000000)
ndim_list = (1, 2)

density_list = (0.1, 0.5)

n_warmup = 1
n_repeat = 10


def get_mean_time(perf):
    cpu_time = perf.cpu_times.mean()
    gpu_time = perf.gpu_times.mean()
    return max(cpu_time, gpu_time) * 1000  # ms


def time_nonzero(a):
    cp_a = cupy.array(a)
    ref = numpy.nonzero(a)
    ret = cupy.nonzero(cp_a)
    for _ref, _ret in zip(ref, ret):
        testing.assert_allclose(_ret, _ref)

    times = []
    perf = repeat(numpy.nonzero, (a,),
                  n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_mean_time(perf))

    perf = repeat(cupy.nonzero, (cp_a,),
                  n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_mean_time(perf))
    return times


def time_argwhere(a):
    cp_a = cupy.array(a)
    ref = numpy.argwhere(a > 0)
    ret = cupy.argwhere(cp_a > 0)
    testing.assert_allclose(ret, ref)
    
    times = []
    perf = repeat(numpy.argwhere, (a > 0,),
                  n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_mean_time(perf))

    perf = repeat(cupy.argwhere, (cp_a > 0,),
                  n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_mean_time(perf))
    return times


def decide_shape(size, ndim):
    shape = []
    for i in range(ndim, 0, -1):
        s = int(math.pow(size, 1 / i) + .5)
        size /= s
        shape.append(s)
    return tuple(shape)


for routine in routine_list:
    print('routine: {}'.format(routine))
    print('shape | density | NumPy | CuPy')
    print('------|---------|-------|-----')
    for size in size_list:
        for ndim in ndim_list:
            for density in density_list:
                shape = decide_shape(size, ndim)
                a = numpy.random.random(shape)
                a[a < 1 - density] = 0
                if routine == 'nonzero':
                    ret = time_nonzero(a)
                elif routine == 'argwhere':
                    ret = time_argwhere(a)
                print('{} | {} | {:.3f} ms | {:.3f} ms'.
                      format(shape, density, ret[0], ret[1]))
    print()
```
</details>